### 2020-08-31 #2615 @danielskatz
👋 @jpsety - I hope you can help me by suggesting some possible reviewers...

If you have any suggestions for potential reviewers, then please mention them here in this thread (without tagging them with an @).

In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
