### 2020-08-31 #2615 @jpsety
Looking through the list its sparse for hardware people, but maybe:
julianstirling
skadio
reality
