### 2020-09-02 #2615 @danielskatz
👋 @julianstirling, @skadio, @reality - Are one or two of you willing to review this submission for JOSS?
