### 2020-09-02 #2615 @julianstirling
I am a hardware person, but I have very little FPGA experience which I think this project needs. [Rowan Hoggarth at Imperial](http://www.imperial.ac.uk/a-z-research/quantum-optics-and-laser-science/people/academic-and-research-staff/) would have the correct experience I would think.
