### 2020-09-03 #2615 @skadio
Not a hardware person here, from Boolean circuits/SAT perspective I can provide a review for this with the exception that the next few weeks is tricky for me.
