### 2020-09-03 #2615 @jpsety
I might have misspoke, a hardware background isn't necessary. Really a Boolean circuits/SAT perspective is what is best.
