### 2020-09-03 #2615 @skadio
Hardware/fpga background would be good to create "interesting" circuits I guess, but, at a second look, this would be a good fit for me. I vaguely remember my Karnaugh maps and logical circuit design.

Quick question:

I am about to submit a paper to JOSS from our research group. Any conflict of interest I should be declaring? I checked the Ethics Guideline with no immediate concern but worth double-checking about serving as a reviewer and in parallel having an (unrelated) submission under review.
