### 2020-09-03 #2615 @julianstirling
Could also try [Frank Seifert](https://www.nist.gov/people/frank-seifert) very experience electrical engineer with FPGA python experience.
