### 2020-09-04 #2615 @danielskatz
@skadio - thanks. There is no conflict between submitting and reviewing. I'll add you as a reviewer, and once I get some info from Rowan, I'll also add him and we can start the review.  Don't worry about the time, if it takes a few weeks, that's ok.  We normally try to get reviews done in 3 weeks, but with Covid, some are taking longer...
