### 2020-09-07 1a0b29b
adding constant wires when writing to verilog
