### 2020-09-08 #2615 @danielskatz
yes, it will appear when we start the review - but you can also see https://joss.readthedocs.io/en/latest/review_checklist.html
