### 2020-09-08 #2646 @whedon
[REVIEW]: CircuitGraph: A Python package for Boolean circuits
**Submitting author:** @jpsety (<a href="http://orcid.org/0000-0002-3775-8638">Joseph Sweeney</a>)
**Repository:** <a href="https://github.com/circuitgraph/circuitgraph" target ="_blank">https://github.com/circuitgraph/circuitgraph</a>
**Version:** v0.0.3
