### 2020-09-08 #2646 @danielskatz
👋 @skadio, @prw99r, @r-a-hoggarth - Thanks again for agreeing to review.  Please read the comments above before starting.

All reviewers have checklists at the top of this thread with the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.
