+++ CHANGELOG.md
@@ -1,0 +1,13 @@
+# Changelog
+All notable changes to this project will be documented in this file.
+
+## [0.0.2] - 2020-09-08
+### Fixed
+- Constant valued nodes are now written to verilog
+- README installation documentation to point to the pypi repo
+
+### Removed
+- pyverilator support, because it breaks installation on python distributions built without tkinter
+
+## [0.0.1] - 2020-08-30
+Initial release
