+++ requirements.txt
@@ -4,5 +4,4 @@ networkx==2.4
 pdoc3==0.8.4
 pyeda==0.28.0
 python-sat==0.1.6.dev4
-PyVerilator==0.1.0
 pyverilog==1.2.1
