### 2020-09-08 6f4f137
Removed pyeda dependency (and removed pyverilator from setup.py)
