+++ circuitgraph/io.py
@@ -5,7 +5,6 @@ import os
 from glob import glob
 import tempfile
 
-from pyeda.parsing import boolexpr
 import pyverilog
 from pyverilog.vparser.parser import VerilogParser
 from pyverilog.vparser import ast as ast_types
