+++ requirements.txt
@@ -2,6 +2,5 @@ codecov==2.1.8
 coverage==5.2.1
 networkx==2.4
 pdoc3==0.8.4
-pyeda==0.28.0
 python-sat==0.1.6.dev4
 pyverilog==1.2.1
