### 2020-09-08 457edbb
Using external link for image in README
Now it will hopefully show up in PyPI
