### 2020-09-08 9e98231
incremented version number in setup.py
