### 2020-09-24 #2646 @prw99r
Just getting into the JOSS way of doing things - learning as I go...

On Thu, Sep 24, 2020 at 5:48 PM Daniel S. Katz <notifications@github.com>
wrote:

> 👋 @skadio <https://github.com/skadio>, @prw99r
> <https://github.com/prw99r>, @r-a-hoggarth
> <https://github.com/r-a-hoggarth> - Just checking in on this again...
>
