### 2020-10-06 #2646 @prw99r
Hi @danielskatz - how do I check off items on the checklist? I'm probably being really dumb, but how the heck do I do that? Thanks, Peter
