### 2020-10-06 #2646 @prw99r
Hi Daniel, I've actually made some progress on looking at the code
(download the repo etc, having a good look at the code etc) - if I can
access the checklist (or there are some commands to do it) then I can
*look* like I'm actually doing something!
Cheers
Peter

On Sun, Oct 4, 2020 at 7:37 PM Daniel S. Katz <notifications@github.com>
wrote:

> It's good to see @r-a-hoggarth <https://github.com/r-a-hoggarth> making
> progress, but I now wonder about @prw99r <https://github.com/prw99r> (no
> boxes checked) and @skadio <https://github.com/skadio> (no changes
> recently). How are things going?
>
