### 2020-10-06 #2646 @prw99r
it says the invitation  has expired?

On Tue, Oct 6, 2020 at 11:59 AM Daniel S. Katz <notifications@github.com>
wrote:

> 👋 @prw99r <https://github.com/prw99r> - you have to accept the
> invitation. See the comment above this one.
>
