### 2020-10-21 #2646 @danielskatz
👋 @skadio, @prw99r, @r-a-hoggarth - Can I again check in with you to see how things are going?  I see a bunch of boxes checked, but some not checked.  For each of you, are the unchecked boxes thing that you just haven't been able to check on yet, or are they items where there is work needed by the submitter, or discussion needed?  If so, please say what these blocking issues are.

Otherwise, can you tell me when you think you will be done with your reviews?
