### 2020-11-02 circuitgraph/circuitgraph#18 @prw99r
Statement of Need Too Narrow and ill defined
The statement of need seems  a bit narrow.

You talk about "tailored to our research needs" which is a bit vague. I don't know (and neither will users) what those research needs are, so please can you define more clearly what the statement of need actually is.

I don't have enough clarity on the requirements, and this would be a good place to really provide specific information.
