### 2020-11-02 #2646 @danielskatz
> A quick update: I looked at the submissions but haven't completed the list yet, hoping to get back to it this weekend.

@skadio - any update?  (I know weekends aren't great for things like this, but just checking...)
