### 2020-11-03 circuitgraph/circuitgraph#18 @jpsety
I've added language that clarifies the specific functionalities that are added.
The research needs are the functionalities provided. (synthesis, sat, composition, interfaces).
The tool itself is not meant to be tied to our research needs, but rather used more broadly. Beyond saying what functionalities our research uses, further explanation isn't relevant.

I've also added a section on the requirements.

Is this helpful?
