### 2020-11-09 #2646 @danielskatz
👋 @prw99r & @skadio - any updates from you?  This review has been going about 2 months, and it would be really good to finish it soon.

I see a bunch of boxes checked, but some not checked. For each of you, are the unchecked boxes thing that you just haven't been able to check on yet, or are they items where there is work needed by the submitter, or discussion needed? If so, please say what these blocking issues are.

Otherwise, can you tell me when you think you will be done with your reviews?
