### 2020-11-10 circuitgraph/circuitgraph#18 @prw99r
Not (much) really, The existing directed graph package in Python provides the basis for description in that format, Verilog provides an industry standard description of logic at a structural and behavioural level, so what do you need that those do not provide? Do we need yet another language? I think you really need to define this properly. Just saying that we need it do these interesting things is not enough I'm afraid. What is the real benefit of having a simple boolean set of descriptors?
