### 2020-11-10 #2646 @prw99r
I've raised a couple of issues with the authors - I hope I'm not being too
picky, but trying to maintain a high standard. I'm not satisfied yet.
Peter

On Mon, Nov 9, 2020 at 1:17 PM Daniel S. Katz <notifications@github.com>
wrote:

> 👋 @prw99r <https://github.com/prw99r> & @skadio
> <https://github.com/skadio> - any updates from you? This review has been
> going about 2 months, and it would be really good to finish it soon.
>
> I see a bunch of boxes checked, but some not checked. For each of you, are
> the unchecked boxes thing that you just haven't been able to check on yet,
> or are they items where there is work needed by the submitter, or
> discussion needed? If so, please say what these blocking issues are.
>
> Otherwise, can you tell me when you think you will be done with your
> reviews?
>
