### 2020-11-10 #2646 @danielskatz
Thanks @prw99r - if you can put https://github.com/openjournals/joss-reviews/issues/2646 in a comment in those issue, then they become visible here and I can see if they are open or closed, which helps me track how this review is going.
