### 2020-11-10 #2646 @prw99r
Daniel,
I've had a think, and I don't think I can approve this for JOSS publication
in its current form.
I took another look at the criteria for publication and guidelines and I
just dont think it satisfies the criteria laid down - it is an OK bit of
coding, as far as it goes, but a major scholarly contribution - no.

I'm not happy about the statement of need, references to prior work, and
also the level of contribution - I just don't see it.

Some useful packages - sure, but major contribution - nope.

It is not about raising specific issues either, that misses the point. it
is the fundamental contribution being asserted - I am not convinced it is
particularly interesting or useful other than to a very limited audience -
hence my comments about the statement of need.

Hope that helps,
Peter

On Mon, Nov 9, 2020 at 1:17 PM Daniel S. Katz <notifications@github.com>
wrote:

> 👋 @prw99r <https://github.com/prw99r> & @skadio
> <https://github.com/skadio> - any updates from you? This review has been
> going about 2 months, and it would be really good to finish it soon.
>
> I see a bunch of boxes checked, but some not checked. For each of you, are
> the unchecked boxes thing that you just haven't been able to check on yet,
> or are they items where there is work needed by the submitter, or
> discussion needed? If so, please say what these blocking issues are.
>
> Otherwise, can you tell me when you think you will be done with your
> reviews?
>
