### 2020-11-10 #2646 @danielskatz
Thanks @prw99r - I don't mean to minimize your concerns; I'm just trying to see where you have raised these issues and what the discussion has been on them.
