### 2020-11-10 #2646 @skadio
@danielskatz  I completed this round of my review. My checklist is updated accordingly and some comments are added below for our authors @jpsety

I hear and share some of the valid points raised by @prw99r. At the moment, I am a bit on the fence with this submission.

Overall, my understanding is that (please correct me if I am wrong) the library provides some I/O related capabilities to read/write circuit instances, an API to create/transform/interact with circuits, ability to do SAT queries.

Under the hood, the bulk of the work is done by the `networkx` library to hold on to that structure and the `pysat` library to answer SAT questions. The standardization/synthesis/analysis are nice contributions but I am not yet convinced these grant an acceptance at JOSS.

On the positive side, the package also ships with some circuit instances which others might find useful (assuming these are interesting instances for the community and are not available in standard formats elsewhere) and the library has the potential to "standardize" circuit operations to some extent. In that sense, one of the major contributions this library can have is "reproducibility" and sharing some common circuit formats --which can accelerate research.

That said, as the authors rightly noted in their submission as well, the library is geared toward benefit the originating lab in the first place. I am not in a good position to fairly judge the applicability/interest from the rest of the community for this tool and/or whether these are gaps are not at all supported by other existing work in this field.

===

- Comment about the codebase: while it seems okay overall but there are some dangling imports, typos, inconsistencies with PEP8 (or just linters in general). The source code can benefit from a quick clean up.

- Codecov and coverage are listed as requirements, but I am not sure if there are necessary strict requirements. The setup.py does not list them either as install requirements so please consider simplifying the installation (or add them as required in setup.py)

- Please explain what's `fanin` and `fanout` for the general reader. It is used in the Composition example without an explanation.

- I couldn't find a contributor policy?

- Have you considered load/save capabilities of Circuit objects? If researchers would like to share or persist the state of their circuit objects if saving/load to bytecode (say pickle files) conforms the integrity. This might be good to check/add.

- @prw99r, I jsut realized that I might have mistakenly checked on/off some of the bullets from your checklist as I was scrolling up and down on the page. I hope I succeeded in retracted my edits but it might worth a quick look.
