### 2020-11-10 circuitgraph/circuitgraph#18 @danielskatz
FYI: from https://github.com/openjournals/joss-reviews/issues/2646
