### 2020-11-10 circuitgraph/circuitgraph#18 @jpsety
Thanks for the comments,
As you say, verilog provides an industry standard description of logic at a structural and behavioral level.
The intent of this library is not to replace verilog; there is already a long list of languages trying to do that (chisel, myhdl, pyrtl, and others). Instead what this library offers is the easy manipulation of _structural_ verilog alongside a long list of other features.

Furthermore, while the library is built upon several existing packages (networkx, pysat), tailoring each of these to represent Boolean circuits is not as trivial as you seem to be suggesting. Yes, at its core, the circuits are represented as a simple set of Boolean descriptors. Built upon that, is a large set of transforms, analysis functions, and generation functions that do not exist in any open-source library, in easy-to-use and prototype language.

A specific example of the power of this library is the following work flow:

- open a list of verilog netlists
- manipulate the netlists algorithmically
- resynthesize the netlists
- run a formal equivalence check between the original and modified netlists

Consider the following pseudo code:
```
#load
circuits = [cg.from_file(n) for n in [circuit names]]

# map some function that uses the library's powerful structural manipulation tools
mod_circuits = circuits.map(arbitrary_transform)

# generic synthesis
syn_mod_circuits = mod_circuits.map(cg.syn)

# check equiv
miter_circuits = [cg.miter(c,cm) for c,cm in zip(circuits,syn_mod_circuits)]
if any(cg.sat('sat':True) for m in miter_circuits):
  print('error')
```

Even with access to industrial priced tools, this process would require swapping between several different tools with inconsistent interfaces, using TCL (a language that is much less accessible than python), and likely an order of magnitude more lines of code.
Without access to those tools, the best solution is probably writing a python script to glue existing open source solutions together. We've done that, and built upon it, adding a wide array of other features.
