### 2020-11-10 circuitgraph/circuitgraph#18 @prw99r
@jpsety - that description is pretty good actually, and something like that in your paper would be really helpful. One key thing that doesn't really come across in the paper is that the core of the work is in the provision of operators and transforms - and this could be described with more clarity. Could you expand on these, with perhaps something like the response above as well? That would improve the description markedly.
