### 2020-11-16 2e354f4
fixes for unused imports etc
