+++ circuitgraph/tests/__init__.py
@@ -1,4 +1,3 @@
 """
 Testing for Python package `circuitgraph`.
 """
-
