### 2020-11-16 #2646 @jpsety
@skadio, Thanks for the feedback, I've addressed most of your comments, but have some further questions/comments.

> Comment about the codebase: while it seems okay overall but there are some dangling imports, typos, inconsistencies with PEP8 (or just linters in general). The source code can benefit from a quick clean up.

I removed the unused imports/some variables I found. Ran the code through the Black linter as specified in  our contribution policy. Fixed typos.

> Codecov and coverage are listed as requirements, but I am not sure if there are necessary strict requirements. The setup.py does not list them either as install requirements so please consider simplifying the installation (or add them as required in setup.py)

Removed these from the requirements.txt, but this has broken our CI testing. Will find a fix.

> Please explain what's fanin and fanout for the general reader. It is used in the Composition example without an explanation.

Added

> I couldn't find a contributor policy?

We have guidelines on the readme of the repo.

> Have you considered load/save capabilities of Circuit objects? If researchers would like to share or persist the state of their circuit objects if saving/load to bytecode (say pickle files) conforms the integrity. This might be good to check/add.

I opened an issue for testing the pickle interface. This is certainly a convenient addition. I believe it should work already, but I'll add some unit tests to be sure.

> Under the hood, the bulk of the work is done by the networkx library to hold on to that structure and the pysat library to answer SAT questions
I think this characterization misses the large amount of work in converting between formats, adding interfaces to generate circuits, transforms that enable SAT calls, etc.

As per the JOSS guidelines:
> Your software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler).
This library makes working with Boolean circuits significantly easier. In only a few lines of code, a user can manipulate circuits in ways that otherwise would take several languages, tools, etc.
See the comment here: [comment](https://github.com/circuitgraph/circuitgraph/issues/18#issuecomment-724761050)
