### 2020-11-16 #2646 @jpsety
@prw99r, I appreciate your feedback. You bring up good points in what can be added. I don't currently have the bandwidth to further edit this paper. As I understand it, the idea of the paper is to give a concise summary of the purpose of the library and show examples of the functionality. I have done that using other accepted projects as examples.
I am confident that the targeted users of this library will be able to use it as is.
