### 2020-11-23 #2646 @skadio
@danielskatz

First of all, I would like to thank @jpsety for the updates and improvements. Most of these comments were not immediate blockers for me but I shared with you in a positive spirit to improve the work for the next/intended users beyond the lead developer. It's nice to see that some are incorporated and others can be added as well.

Without going into a discussion of assessing the exact amount of work involved here, the author highlights an important aspect that I also agree with: "_the large amount of work in converting between formats, adding interfaces to generate circuits, transforms that enable SAT calls_".

I consider this as a valuable contribution since it can help bring consistency across different research applications and for "reproducibility" purposes. In light of that, I don't see an immediate reason why should the research community be barred from learning about this tool and benefit from its functionality. Hence, I would be happy to accept this submission with minor edits from the list above, if any remaining.

Best,
Serdar
