### 2020-11-23 #2646 @prw99r
I agree with the highlighted contribution that @skadio notes above - I
think that @jpsety has gone partially down the road of dealing with my
concerns and added some useful clarifications. The purpose of our comments
has always been to provide positive criticism to improve the software and
the paper. It can be tough to take sometimes, especially if there is a lot
of time and effort already invested in the work, but it is worth taking the
time to do it right.

One nice aspect of this approach (JOSS) is that the issues are there for
all to see, and therefore it is left to the reader to make their own mind
up as to whether the paper (and ultimately the software) will do what they
need it to do, and if the paper is unclear, that they can ask the authors
to clarify further. The last thing I want to do, personally, is get in the
way of others using the library. In many respects that will be a/ the proof
that it is any good and b/ provide useful wider feedback.

I stand by my review comments, JOSS is a Journal after all, albeit in a
different format, and I don't see anything wrong with being rigorous and
demanding, However I do not have a problem with the library being released
for use by the design community, if that is the consensus, and therefore I
am happy to accept the submission now.

Peter
