### 2020-11-23 #2646 @skadio
Revised the Checklist, no need for opening source repo issues from my side.
