### 2020-11-29 #2646 @danielskatz
👋 @prw99r - Can I ask that check off any remaining items in your checklist in [this case](https://github.com/openjournals/joss-reviews/issues/2646#issuecomment-732321669)? If there's anything that you feel you can't check off, let's talk about it here so that it's clear what is needed for acceptance (vs future enhancements that would be good but are not needed for acceptance)

Also, please feel free to open any needed issues in [the source repo](https://github.com/circuitgraph/circuitgraph) as that will be more of a record of potential future enhancement than this review issue, which will close (but remain available) after the review finishes.
