### 2020-12-01 #2646 @prw99r
@danielskatz I have checked off my list now - my comments were for the paper more than the repo, so we are good to go.
