### 2020-12-01 #2646 @danielskatz
Ok!  Thanks very much to everyone (@jpsety, @skadio, @prw99r, @r-a-hoggarth) for working through this process.  I think it's improved the software, and led to a good JOSS paper.
