### 2020-12-01 #2646 @danielskatz
👋 @jpsety - At this point could you:
- [ ] Proofread the paper, and make sure you are happy with it (which I will also do after this part of the process)
