### 2020-12-03 #2646 @jpsety
I've proofread the paper and followed the above instructions.

Tagged release:
https://github.com/circuitgraph/circuitgraph/releases/tag/v0.0.3

DOI:
10.5281/zenodo.4302741
