### 2020-12-04 #2646 @danielskatz
👋 @jpsety - some small final changes are needed in the paper and bib - please merge https://github.com/circuitgraph/circuitgraph/pull/22 or let me know what you disagree with.
