### 2020-12-04 c12488b
Update paper.bib
sorry this didn't make it into the previous PR for some reason
