### 2020-12-04 #2646 @danielskatz
sorry - one more change which I thought was in the PR but apparently wasn't - https://github.com/circuitgraph/circuitgraph/pull/23
