### 2020-12-04 #2646 @danielskatz
@jpsety - sorry, I missed a closing paren in the SciPy2008 bib entry - can you add it and then let me know - this really will be the last change...
