+++ docs/paper.bib
@@ -20,7 +20,7 @@
 @InProceedings{SciPyProceedings_11,
 	  author =       {Aric A. Hagberg and Daniel A. Schult and Pieter J. Swart},
 	    title =        {Exploring Network Structure, Dynamics, and Function using NetworkX},
-		  booktitle =   {{Proceedings of the 7th Python in Science Conference (SciPy2008}},
+		  booktitle =   {{Proceedings of the 7th Python in Science Conference (SciPy2008)}},
 		    pages =     {11 - 15},
 			  address = {Pasadena, CA USA},
 			    year =      {2008},
